requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
slack-bolt==1.18.0
slack-sdk==3.23.0
APScheduler==3.10.4
python-dotenv==1.0.0
gunicorn==21.2.0
